searchAIID-service/pyproject.toml
[project]
name = "searchAIID-service"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "dannhanong", email = "thannong0512@gmail.com"},
]
dependencies = [
    "fastapi>=0.115.12", 
    "uvicorn>=0.34.2", 
    "transformers>=4.51.3", 
    "opencv-python>=4.11.0.86", 
    "python-dotenv>=1.1.0", 
    "torch>=2.7.0", 
    "langchain>=0.3.25", 
    "pillow>=11.2.1", 
    "python-multipart>=0.0.20", 
    "torchvision>=0.22.0", 
    "pydantic-settings>=2.9.1", 
    "einops>=0.8.1", 
    "timm>=1.0.15", 
    "pymongo>=4.12.1", 
    "scikit-learn>=1.6.1", 
    "pandas>=2.2.3", 
    "numpy>=2.2.5", 
    "pyjwt>=2.10.1", 
    "confluent-kafka>=2.10.0",
]
requires-python = ">=3.12"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true
